refactoring reply function
